Add weight quantization implementation and documentation
- Add qwen3_patch_weightQuant.py: New weight quantization approach for Qwen3 attention
- Add CLAUDE.md: Comprehensive documentation for the FP4 attention quantization codebase
- Modified llama_patch.py: Include existing changes from cleaned_code branch

This branch introduces weight quantization as an alternative to activation quantization,
focusing on quantizing projection matrices (Q, K, V, O) rather than intermediate activations.

+++ llama_patch.py
@@ -613,5 +613,5 @@ def llama_fp4_attention_forward(
         attn_output = attn_output.to(self.dequant_dtype)
 
     attn_output = attn_output.reshape(*input_shape, -1).contiguous()
-    attn_output = self.o_proj(attn_output)
+    attn_output = self.o_proj(attn_output.to(self.o_proj.weight.dtype))
     return attn_output, attn_weights
